Create securitas platform instead of a separate sensor and switch.
